Big commit, oops.
The biggest part of the commit that I remember is adding the bar plot, of which I got from a website credited in the source code and in the readme. I spent some time modifying it and adding a statistic class which also needs to be finished getting fixed but currently I am adding a search filter.